add apriori

+++ movie.py
@@ -1,17 +1,18 @@
-from flask import render_template, jsonify, request, redirect, url_for, Response
+import datetime
+import json
+import os
+
+from flask import (Response, jsonify, make_response, redirect, render_template,
+                   request, send_file, send_from_directory, url_for)
+from flask_login import (LoginManager, current_user, login_required,
+                         login_user, logout_user)
 from sqlalchemy import func
+from werkzeug.utils import secure_filename
 
 import models
-from forms import LoginForm
 from settings import app, db
-from flask_login import login_user, login_required
-from flask_login import LoginManager, current_user
-from flask_login import logout_user
-from flask_wtf.csrf import CsrfProtect
-import os
-import json
-from flask import send_from_directory
 
+VIDEO_SAVE_PATH = 'video/'
 PAGE_SIZE = 10
 # app.secret_key = os.urandom(24)
 
@@ -20,11 +21,6 @@ login_manager = LoginManager()
 # login_manager.session_protection = 'strong'
 login_manager.login_view = 'login'
 login_manager.init_app(app=app)
-
-
-# csrf protection
-# csrf = CsrfProtect()
-# csrf.init_app(app)
 
 
 @login_manager.user_loader
@@ -42,6 +38,8 @@ def login():
         password = request.form['password']
         user = models.User.query.filter(models.User.account == account).first()
         if user is not None:
+            if user.is_freeze == True:
+                return jsonify('账号被冻结，请联系管理员解封')
             if user.verify_password(password):
                 # user = models.User(account)
                 login_user(user)
@@ -78,8 +76,9 @@ def register():
 @app.route('/main')
 @login_required
 def main():
+    print(request.remote_addr)
     return render_template(
-        'main.html')
+        'movieIndex.html')
 
 
 # ...
@@ -97,7 +96,7 @@ def get_all_users():
     for i in users:
         res.append({'id': i.id, 'username': i.username, 'account': i.account,
                     'email': i.email, 'register_date': str(i.register_date),
-                    'is_admin': i.is_admin, 'phone': i.phone})
+                    'is_admin': i.is_admin, 'phone': i.phone, 'is_freeze': i.is_freeze})
     return jsonify(res)
 
 
@@ -105,7 +104,6 @@ def get_all_users():
 @login_required
 def give_admin():
     id = request.args.get('id')
-    print(id)
     user = models.User.query.get(id)
     if user is None:
         return jsonify('user not exists')
@@ -170,6 +168,32 @@ def delete_user():
     return 'true'
 
 
+@app.route('/freeze_user')
+@login_required
+def freeze_user():
+    id = request.args.get('id')
+    u = models.User.query.get(id)
+
+    if u is None:
+        return 'user_not_exists'
+    u.is_freeze = True
+    db.session.commit()
+    return 'success'
+
+
+@app.route('/release_freeze')
+@login_required
+def release_freeze():
+    id = request.args.get('id')
+    u = models.User.query.get(id)
+
+    if u is None:
+        return 'success'
+    u.is_freeze = False
+    db.session.commit()
+    return 'success'
+
+
 @app.route('/pub_news', methods=['POST'])
 @login_required
 def publish_news():
@@ -214,10 +238,65 @@ def get_all_movies():
     return jsonify(res)
 
 
+@app.route('/update_movie', methods=['POST'])
+@login_required
+def update_movie():
+    name = request.form['name']
+    id = request.form['id']
+    actor = request.form['actor']
+    director = request.form['director']
+    genre = request.form['genre']
+
+    movie = models.Movie.query.get(int(id))
+    if movie is None:
+        return jsonify('success')
+
+    movie.name = name if name is not None else movie.name
+    movie.actor = actor if actor is not None else movie.actor
+    movie.director = director if director is not None else movie.director
+    movie.genre = genre if genre is not None else movie.genre
+
+    db.session.commit()
+
+    return jsonify('success')
+
+
 @app.route('/movies_count')
 def get_movies_count():
     count = models.Movie.query.count()
     return jsonify(count)
+
+
+@app.route('/del_movie')
+def del_movie():
+    id = request.args.get('id')
+    movie = models.Movie.query.get(id)
+    if movie is None:
+        return jsonify('success')
+
+    try:
+        # 删除电影，同时删除电影之下的评论，收藏下的电影，movie_cat_re,movie_eva,user_collection
+        # 删除评论
+        evas = models.MovieEva.query.filter(models.MovieEva.movie_id == id).all()
+        for i in evas:
+            db.session.delete(i)
+        # 删除包含该分类的电影
+        cat = models.MovieCatRe.query.filter(models.MovieCatRe.movie_id == id).all()
+        for c in cat:
+            db.session.delete(c)
+
+        # 删除用户收藏的电影
+        uc = models.UserCollection.query.filter(models.UserCollection.movie_id == id).all()
+        for u in uc:
+            db.session.delete(u)
+        # 删除电影本身
+        db.session.delete(movie)
+
+        db.session.commit()
+
+    except:
+        return jsonify('error')
+    return jsonify('success')
 
 
 @app.route('/page_movies')
@@ -233,13 +312,10 @@ def get_page_movie():
     current_page = int(request.args.get('curr_page'))
     sort_key = request.args.get('sort_key')
 
-    search_words = request.args.get('search_words')
-
-    if search_words is not None:
-        print(search_words)
+    search_words = str(request.args.get('search_words'))
+    if search_words != '_none':
         movies = models.Movie.query.order_by(db.desc(s.get(sort_key, 'id'))) \
-            .filter(models.Movie.name.like(''.join(['%', search_words, '%']))).limit(PAGE_SIZE).offset(
-            (current_page - 1) * PAGE_SIZE)
+            .filter(models.Movie.name.like(''.join(['%', search_words, '%']))).all()
     else:
         movies = models.Movie.query.order_by(db.desc(s.get(sort_key, 'id'))).limit(PAGE_SIZE).offset(
             (current_page - 1) * PAGE_SIZE)
@@ -253,11 +329,30 @@ def get_page_movie():
         'genre': m.genre,
         'actor': m.actor,
         'director': m.director,
-        'score': m.score,
-        'views': m.views,
+        'score': str(m.score),
+        'views': str(m.views),
         'video_path': m.video_path
     } for m in movies]
+    return jsonify(res)
 
+
+@app.route('/search_movies')
+def search_movies():
+    s = request.args.get('search_words')
+    if s is None:
+        return jsonify([])
+    movies = models.Movie.query.order_by(db.desc(models.Movie.name)) \
+        .filter(models.Movie.name.like(''.join(['%', s, '%']))).all()
+    res = [{
+        'id': m.id,
+        'name': m.name,
+        'genre': m.genre,
+        'actor': m.actor,
+        'director': m.director,
+        'score': str(m.score),
+        'views': str(m.views),
+        'video_path': m.video_path
+    } for m in movies]
     return jsonify(res)
 
 
@@ -280,6 +375,8 @@ def get_all_guests():
 def del_guest():
     id = request.args.get('id')
     g = models.GuestBook.query.filter(id == models.GuestBook.id).first()
+    if g is None:
+        return jsonify('success')
     db.session.delete(g)
     db.session.commit()
     return jsonify('success')
@@ -387,6 +484,41 @@ def del_movie_from_category():
     return jsonify('success')
 
 
+@app.route('/update_category', methods=['POST'])
+@login_required
+def update_category():
+    id = request.form['id']
+    name = request.form['name']
+    desc = request.form['desc']
+
+    c = models.MovieCategory.query.get(int(id))
+    if c is None:
+        return jsonify('error')
+    c.category = name
+    c.desc = desc
+    db.session.commit()
+    return jsonify('success')
+
+
+@app.route('/del_category')
+def delete_category():
+    id = request.args.get('id')
+
+    c = models.MovieCategory.query.get(int(id))
+
+    if c is None:
+        return jsonify('error')
+    # 先删除下面收录的电影
+
+    re = models.MovieCatRe.query.filter(models.MovieCatRe.movie_cat_id == id).all()
+    for r in re:
+        db.session.delete(r)
+    # 再删除该类别
+    db.session.delete(c)
+    db.session.commit()
+    return jsonify('success')
+
+
 @app.route('/movie_details_page/<movie_id>', methods=['GET'])
 def movie_details_page(movie_id):
     movie = models.Movie.query.get(int(movie_id))
@@ -397,8 +529,9 @@ def movie_details_page(movie_id):
     return render_template('movieDetail.html', movie=movie)
 
 
-@app.route('/movie_details/<movie_id>')
-def movie_details(movie_id):
+@app.route('/movie_details')
+def movie_details():
+    movie_id = request.args.get('movie_id')
     movie = models.Movie.query.get(int(movie_id))
     return jsonify(
         {'id': movie.id,
@@ -408,9 +541,20 @@ def movie_details(movie_id):
          'director': movie.director,
          'score': movie.score,
          'views': movie.views,
-         'video_path': movie.video_path
+         'video_path': movie.video_path,
+         'collect_num': movie.collect_num,
+         'eva_number': movie.eva_num
          }
     )
+
+
+@app.route('/movie_video/<movie_id>')
+def get_movie_video(movie_id):
+    movie = models.Movie.query.get(int(movie_id))
+    if movie is not None:
+        return send_file(movie.video_path[1:])
+    else:
+        return 'error'
 
 
 @app.route('/movie_comments')
@@ -441,17 +585,21 @@ def add_new_comments():
     me.score = score
     me.movie_id = movie_id
     me.user_id = curr_user_id
-    # 更新 movie 表中的平均得分
-    avg_score = db.session.query(func.avg(models.MovieEva.score)
-                                 .label('average')).filter(models.MovieEva.movie_id == movie_id).first()[0]
-    movie = models.Movie.query.get(movie_id)
-    if movie is not None:
-        movie.score = avg_score
 
     db.session.add(me)
     db.session.commit()
+    # 更新 movie 表中的平均得分
+    avg_score = db.session.query(func.avg(models.MovieEva.score)
+                                 .label('average')).filter(models.MovieEva.movie_id == movie_id).first()[0]
+    print(avg_score)
+    movie = models.Movie.query.get(movie_id)
+    if movie is not None:
+        movie.score = avg_score
+        movie.eva_num += 1
 
-    return jsonify('success')
+    db.session.commit()
+
+    return jsonify({'res': 'success', 'score': avg_score})
 
 
 @app.route('/questionnaire')
@@ -488,8 +636,11 @@ def get_statistics():
     score = db.session.query(func.avg(models.Qa.score).label('average')).first()[0]
     from_where = db.session.query(models.Qa.from_where, func.count('*')) \
         .group_by(models.Qa.from_where).all()
+    from_where = {i[0]: i[1] for i in from_where}
+
     favorite_genre = db.session.query(models.Qa.favorite_genre, func.count('*')) \
         .group_by(models.Qa.favorite_genre).all()
+    favorite_genre = {i[0]: i[1] for i in favorite_genre}
     res = {
         'movie_count': movie_count,
         'user_count': user_count,
@@ -505,7 +656,7 @@ def get_statistics():
 
 @app.route('/qa')
 def get_all_qa():
-    qa = models.Qa.query.order_by(db.desc(models.Qa.submit_date)).all()
+    qa = models.Qa.query.order_by(db.desc(models.Qa.submit_date)).limit(50)
     return jsonify([{'score': i.score, 'submit_date': str(i.submit_date),
                      'favorite_genre': i.favorite_genre,
                      'sex': i.sex, 'from_where': i.from_where} for i in qa])
@@ -524,8 +675,12 @@ def collect_movie():
 
     # 已经存在，直接返回
     if len(is_exist) != 0:
-        return jsonify('success')
+        return jsonify('collected')
 
+    # 更新 movie 表中 collect_num 字段
+    m = models.Movie.query.get(int(movie_id))
+    if m is not None:
+        m.collect_num += 1
     c = models.UserCollection(movie_id, user_id)
     c.mark = mark
     db.session.add(c)
@@ -533,16 +688,92 @@ def collect_movie():
     return jsonify('success')
 
 
-@app.route('/my_collections')
+@app.route('/personal_collect')
 @login_required
-def collections_page():
-    return render_template('myCollection.html')
+def get_all_my_collections():
+    user = current_user
+    c = models.UserCollection.query.order_by(db.desc(models.UserCollection.collect_date)).filter(
+        models.UserCollection.user_id == user.id).all()
+    res = [{
+        'id': i.id,
+        'movie_id': i.movie.id,
+        'movie_name': i.movie.name,
+        'collect_date': str(i.collect_date),
+        'mark': i.mark}
+        for i in c
+    ]
+    return jsonify(res)
+
+
+@app.route('/del_collection')
+@login_required
+def del_collection():
+    id = request.args.get('id')
+    if id is None:
+        return 'error'
+    c = models.UserCollection.query.get(int(id))
+    if c is None:
+        return 'success'
+    db.session.delete(c)
+
+    db.session.commit()
+    return jsonify('success')
+
+
+@app.route('/upload_video/<movie_id>', methods=['POST', 'GET'])
+def upload(movie_id):
+    if movie_id is None:
+        return jsonify('error')
+    movie = models.Movie.query.get(int(movie_id))
+    if movie is None:
+        return jsonify('error')
+
+    f = request.files['file']
+    basepath = os.path.dirname(__file__)  # 当前文件所在路径
+    filename = str(datetime.datetime.now()) + f.filename
+    upload_path = os.path.join(basepath, 'static/video', secure_filename(filename))  # 注意：没有的文件夹一定要先创建，不然会提示没有该路径
+    f.save(upload_path)
+
+    movie.video_path = '/static/video/' + secure_filename(filename)
+
+    db.session.commit()
+    return jsonify({'res': 'success', 'video_path': movie.video_path})
+
+
+@app.route('/news_detail/<id>')
+def news_details(id):
+    news = models.News.query.get(int(id))
+    return render_template('newsDetail.html', news=news)
 
 
 @app.route('/guest')
-@login_required
 def guest_page():
     return render_template('guestBook.html')
+
+
+@app.route('/movie_index')
+def movie_index():
+    return render_template('movieIndex.html')
+
+
+@app.route('/news_page')
+def news_page():
+    return render_template('news.html')
+
+
+@app.route('/user_manage')
+def user_manage():
+    return render_template('user_manage.html')
+
+
+@app.route('/my_collection')
+def collection_page():
+    return render_template('myCollection.html')
+
+
+@app.route('/test')
+def just_test():
+    return render_template('test.html')
 
 
 if __name__ == '__main__':
